Iss19 (#21)
* Create scripts_readme.yml

Creating a workflow to reflect the "Available Datasets" material in the ./scripts/README.md file of the repository. For now, the Action is set to be triggered manually.

* Create scripts_readme.yml

Creating a workflow to reflect the "Available Datasets" material in the ./scripts/README.md file of the repository. For now, the Action is set to be triggered manually.

* modifying the script to trigger whenever there is a push on main branch

* Update scripts_readme.yml

* Adding manual trigger as well