Adding notes for where `logs` are created
This raises awareness about various `logs` generated during the process.
This has been felt as many users are struggling to investigate various
issues that may arise during the processing of various datasets.

The issues could be isolated to the scripts themselves which is a rare
occasion, or could be related to the hardware infrastructure of the
processing unit.

Signed-off-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>