Ouranos ESPO-G6-R2 script + new capablities
This script introduces new features to the tool, including the
capability to process the climate datasets, including those consisting
of multiple models, submodels (those with specific configuration sets),
ensemble members, and multiple scenarios (SSPs). The parent calling
script is in charge of parallelization scheme, if needed.

With this script, a few issues related to the current deficiencies of
datatool could be resolved simultaneously.

Signed-off-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>