Adding new parallelization schemes
Multi parallelization schemes are added, so the package not only submit
array jobs based on the given date range and the chunk schemes, but also
considers submitting jobs based on various models, ensemble members, and
scenarios. These new parallelization schemes mostly applies to climate
datasets, but not necessarily.

This commit aims to save time for the user and fasten the processing
time for datasets.

This commit resolves issue #25 on remote GitHub hosting repository.
Furthermore, it adds the ESPO dataset to the list of datasets as well.

Moreover, a new option is implement to show the list of currently
available datasets to the users.

Signed-off-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>