Updates to the documentations
The help message has been trimmed to provide more information to the
users. This include values provided to the `--lon-lims` that must be
within the [-180, +180] limits. This has not been mentioned before to
the users and could have provided confusion, as there are multiple
methods to describe longitudes.

Furthermore, the list of datasets on the main page of the repository has
been updated to reflect the most up-to-date list.

Signed-off-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>