Assuring RDRS ends properly
In this commit, two concerns are addressed: 1) for small domains some
times, the process ends without actually doing the job and 2) using ncks
and index-based hyperslab to extract a spatial subset.

For the former concern, a `until ... do` loop is used to assure each
process on each file is finished properly. Furthermore, proper error
messages are thrown at the user. For a simply case of `Bow River at
Banff`, where the final NetCDF files are in the magnitute of a few
kilobytes, the script was not behaving properly before. This behaviour
is as expected with this commit.

For the latter concern, the `ncks` is used to avoid using `cdo`. `ncks`
proves to be more stable and reliable.

Signed-off-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>