Addressing last date error with RDRS script
The RDRS script now retries as many times as necessary to produce the
final files. However, this used to also valid when the file did not
exists. This commit resolves this issue, and assure if a date beyond the
last file available is given to the script, this is corrected.

Signed-off-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>