Assuring the spatial domain is fully covered
In this commit, a value of 0.1 degrees is added/subtracted to/from the
minimum and maximum values of latitude and longitude extents. This is to
assure that the spatial domain of the requested area is fully covered in
the subsetted dataset files. Previously, the extents were not accurately
covered.

The value of 0.1 degress reflect the resolution of RDRSv2.1.

Signed-off-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>