Considering future runs of GWF WRF-CONUS II product
The future runs, unfortunately, follow a different file/folder structure
and they needed to be handled a bit differently than the historical
files which are all in `.tar` format.

This PR adds the capability to process this period of data for this
product to the script. Minimal adjustments were done to make the script
almost intact.

Reported-by: Logan Fang <xif382@usask.ca>
Signed-off-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>