Taking care of time-steps in the middle of the hour
In cases, the middle hour time-step was printed as HH:29:599999 so it
cause issues with time shift, as 29 != 30 (30 was the assumption
previosuly).

Furthermore, the last time step of subsetting is assumed to be HH:45:00
so as to no miss time-steps where the value is set to HH:30:00001
(slightly past HH:30:00).

Reported-by: Mohamed Elshamy <mohamed.elshamy@usask.ca>
Signed-off-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>